requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
appnope==0.1.3
asttokens==2.4.0
async-timeout==4.0.3
attrs==23.1.0
backcall==0.2.0
Brotli==1.1.0
certifi==2023.7.22
charset-normalizer==3.2.0
comm==0.1.4
cramjam==2.7.0
DateTime==5.2
debugpy==1.8.0
decorator==5.1.1
executing==1.2.0
fastparquet==2023.8.0
frozenlist==1.4.0
fsspec==2023.9.1
idna==3.4
inflate64==0.3.1
ipykernel==6.25.2
ipython==8.15.0
jedi==0.19.0
jupyter_client==8.3.1
jupyter_core==5.3.1
matplotlib-inline==0.1.6
multidict==6.0.4
multivolumefile==0.2.3
nest-asyncio==1.5.8
numpy==1.26.0
openai==0.28.0
packaging==23.1
pandas==2.1.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==3.10.0
prompt-toolkit==3.0.39
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
py7zr==0.20.6
pyarrow==13.0.0
pybcj==1.0.1
pycryptodomex==3.19.0
Pygments==2.16.1
pyppmd==1.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
pyzmq==25.1.1
pyzstd==0.15.9
requests==2.31.0
six==1.16.0
stack-data==0.6.2
texttable==1.6.7
toml==0.10.2
tornado==6.3.3
tqdm==4.66.1
traitlets==5.10.0
tzdata==2023.3
urllib3==2.0.5
wcwidth==0.2.6
yarl==1.9.2
zope.interface==6.0
